Removed and updated the routes to point to revisions.service.ts. Opted to choose the async createRevision function over the series of then statements. Got rid of unused parameter for deleteRevision in the service. Updated all function calls.
