+++ src/app/admin/core/unrelease.service.ts
@@ -10,25 +10,6 @@ import { REVISION_ROUTES } from '../../core/learning-object-module/revisions/rev
 })
 export class UnreleaseService {
   constructor(private http: HttpClient) {}
-
-  /**
-   * Deletes a revision of a learning object. This is designed to allow an editor to create a new
-   * revision when it is necessary for the editorial process to continue.
-   *
-   * @param username username of the author
-   * @param cuid cuid of the learning object
-   * @returns
-   */
-  deleteRevision(username: string, cuid: string, version: number) {
-    return this.http
-      .delete(REVISION_ROUTES.DELETE_REVISION(cuid, version), {
-        withCredentials: true,
-        responseType: 'text',
-      })
-      .pipe(catchError(this.handleError))
-      .toPromise();
-  }
-
   /**
    * Generic error-handling function for errors through from the HttpClient module
    *
