+++ src/app/cube/learning-object.service.ts
@@ -10,8 +10,6 @@ import {
 import { Query } from '../interfaces/query';
 
 import { SEARCH_ROUTES } from 'app/core/learning-object-module/search/search.routes';
-import * as querystring from 'querystring';
-import { REVISION_ROUTES } from '../core/learning-object-module/revisions/revisions.routes';
 import { USER_ROUTES } from '../core/user-module/user.routes';
 
 // TODO: move to core module
@@ -132,20 +130,6 @@ export class LearningObjectService {
       .then((val: any) => {
         return val.map((l) => new LearningObject(l));
       });
-  }
-
-  /**
-   * Creates a Revision of an existing learning object
-   *
-   * @param cuid the CUID of the learning object to create a revision of
-   */
-  async createRevision(cuid: string): Promise<any> {
-    const route = REVISION_ROUTES.CREATE_REVISION(cuid);
-    const response = await this.http
-      .post(route, {}, { withCredentials: true })
-      .pipe(catchError(this.handleError))
-      .toPromise();
-    return response;
   }
 
   private handleError(error: HttpErrorResponse) {
